+++ src/codegen/gql2ts.ts
@@ -1,3 +1,4 @@
+import { toError } from "../util/toError";
 import {
   CodeGenerator,
   CodeGeneratorOptions,
@@ -20,7 +21,7 @@ export async function gql2ts(options: Gql2TsOptions) {
     const codeGenerator = new CodeGenerator(options);
     await codeGenerator.cli(options);
   } catch (err) {
-    logger.error(err.stack);
+    logger.error(toError(err));
     if (!options.noExit) process.exit(1);
   }
 }